backend/pyproject.toml
[tool.poetry]
name = "learnLab-backend"
version = "0.1.0"
description = "LearnLab Backend"
authors = ["Surya <rebbapragada.s@northeastern.edu>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
langchain = "^0.3.7"
langchain-core = "^0.3.18"
langgraph = "^0.2.48"
langchain-community = "^0.3.7"
langchain-groq = "^0.2.1"
langgraph-checkpoint = "^2.0.4"
langgraph-checkpoint-sqlite = "^2.0.1"
duckduckgo-search = "^6.3.4"
arxiv = "^2.1.3"
numexpr = "^2.10.1"
python-dotenv = "^1.0.1"
pydantic = {extras = ["email"], version = "^2.9.2"}
langchain-openai = "^0.2.10"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}  # Added [standard] for WebSocket support
websockets = "12.0"  # Added explicit websockets dependency
sqlalchemy = "^2.0.28"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
boto3 = "^1.34.69"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
aiofiles = "^23.1.0"  
pydub = "^0.25.1"     
ffmpeg-python = "^0.2.0" 
semantic-router = "^0.0.72"
pypdf2 = "^3.0.1"
pinecone = "^5.4.2"
debugpy = "^1.6.7"
upstash-vector = "^0.6.0"
logger = "^1.4"
google-auth = "^2.37.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.155.0"
ratelimit = "^2.2.1"
backoff = "^2.2.1"
markdown = "^3.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "."
venv = ".venv"